feat: Implement duel submission tab with stats tracking
- Created DuelRecords page with sidebar showing real-time stats
- Built duel submission form with searchable deck dropdowns
- Added coin flip, turn order, and match result selection
- Integrated decimal rating support for rated/duelist cup modes
- Implemented color-coded win/loss record display (green/red)
- Connected submission to backend API with proper stats updates
- Fixed RPC functions to support decimal point changes
- Optimized scrollbar visibility across different tabs
- Added real-time stat refresh after duel submission

+++ database/rpc-functions.sql
@@ -28,7 +28,7 @@ CREATE OR REPLACE FUNCTION update_rated_stats(
   p_session_id INTEGER,
   p_user_id VARCHAR(20),
   p_result VARCHAR(10),
-  p_point_change INTEGER
+  p_point_change DECIMAL(10,2)
 )
 RETURNS VOID AS $$
 BEGIN
@@ -48,7 +48,7 @@ CREATE OR REPLACE FUNCTION update_duelist_cup_stats(
   p_session_id INTEGER,
   p_user_id VARCHAR(20),
   p_result VARCHAR(10),
-  p_point_change INTEGER
+  p_point_change DECIMAL(10,2)
 )
 RETURNS VOID AS $$
 BEGIN
